feat: add Docker build and run scripts
The package.json file has been modified to include new scripts for building and running the Docker image. This will allow for easier deployment and testing of the application in a containerized environment.

Recent user commits and repository commits are not directly related to the code changes, so they are not considered in the commit message.