Feat/docker backend (#1)
* create first basic dockerfile

* remove app from the docker image

* feat: add Docker build and run scripts

The package.json file has been modified to include new scripts for building and running the Docker image. This will allow for easier deployment and testing of the application in a containerized environment.

Recent user commits and repository commits are not directly related to the code changes, so they are not considered in the commit message.

* chore: update Dockerfile to include package installation steps

* chore: adds a base express server and runs it in docker-run

---------

Co-authored-by: Adi De Masi <adi.demasi@gmail.com>
Co-authored-by: Sergej Brazdeikis <sergej.brazdeikis@swissmarketplace.group>